+++ internal/cli/parse.go
@@ -166,7 +166,7 @@ func ParseArgs(args []string) (run bool, arguments Arguments) {
 
 	err := doParseProgramArguments(args[1:], &arguments)
 	if err != nil {
-		logger.Error(err)
+		logger.Fatalf("An error occurred while parsing arguments: %s", err)
 		return false, arguments
 	}