+++ internal/cli/arg_context.go
@@ -10,9 +10,6 @@ const (
 	// The context where arguments are interpreted as input files.
 	contextDefault argContext = iota
 
-	// The context where arguments are interpreted as a configuration file.
-	contextConfigFile
-
 	// The context where arguments are interpreted as mapping files.
 	contextMapFile
 
@@ -24,7 +21,6 @@ const (
 func (context argContext) String() string {
 	names := []string{
 		"Unknown",
-		fmt.Sprintf("%s/%s", configOption.name, configOption.alias),
 		fmt.Sprintf("%s/%s", mapfileOption.name, mapfileOption.alias),
 		fmt.Sprintf("%s/%s", mappingOption.name, mappingOption.alias),
 	}