+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning].
 ### Bug Fixes
 
 - Fix bug preventing the usage of map files without file extension. ([#101])
+- Fix logs of incorrect level being logged. ([#107], [#108])
 
 ## [0.5.0-beta] - 2020-07-24
 
@@ -118,3 +119,5 @@ Versioning].
 [#98]: https://github.com/ericcornelissen/wordrow/pull/98
 [#99]: https://github.com/ericcornelissen/wordrow/pull/99
 [#101]: https://github.com/ericcornelissen/wordrow/pull/101
+[#107]: https://github.com/ericcornelissen/wordrow/pull/107
+[#108]: https://github.com/ericcornelissen/wordrow/pull/108