+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning].
 ### Bug Fixes
 
 - Fix bug preventing the usage of map files without file extension. ([#101])
+- Fix logs of incorrect level being logged. ([#107], [#108])
 
 ### Miscellaneous
 
@@ -128,3 +129,4 @@ Versioning].
 [#101]: https://github.com/ericcornelissen/wordrow/pull/101
 [#102]: https://github.com/ericcornelissen/wordrow/pull/102
 [#107]: https://github.com/ericcornelissen/wordrow/pull/107
+[#108]: https://github.com/ericcornelissen/wordrow/pull/108