+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Versioning].
 - Fix bug preventing the usage of map files without file extension. ([#101])
 - Fix logs of incorrect level being logged. ([#107], [#108])
 - Fix crash due to empty values in the `--map` option. ([#110])
+- Fix mistakes in the program help message. ([#114])
 
 ### Miscellaneous
 
@@ -132,3 +133,4 @@ Versioning].
 [#107]: https://github.com/ericcornelissen/wordrow/pull/107
 [#108]: https://github.com/ericcornelissen/wordrow/pull/108
 [#110]: https://github.com/ericcornelissen/wordrow/pull/110
+[#114]: https://github.com/ericcornelissen/wordrow/pull/114